Refactored collapse and PDF render
