+++ app/message-attachments/client/messageAttachment.html
@@ -62,11 +62,7 @@
 					{{else}}
 						{{title}}
 					{{/if}}
-					{{#if collapsed}}
-						<span class="collapse-switch icon-right-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
-					{{else}}
-						<span class="collapse-switch icon-down-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
-					{{/if}}
+					{{> collapseArrow collapsedMedia=collapsedMediaVar}}
 				</div>
 			{{/if}}
