--- app/oembed/client/oembedAudioWidget.js
@@ -1,7 +1,0 @@
-import { Template } from 'meteor/templating';
-
-Template.oembedAudioWidget.helpers({
-	collapsed() {
-		return this.collapsedMedia;
-	},
-});
