--- app/oembed/client/oembedFrameWidget.js
@@ -1,7 +1,0 @@
-import { Template } from 'meteor/templating';
-
-Template.oembedFrameWidget.helpers({
-	collapsed() {
-		return this.collapsedMedia;
-	},
-});
