+++ app/oembed/client/oembedImageWidget.html
@@ -1,9 +1,7 @@
 <template name="oembedImageWidget">
 	{{#if parsedUrl}}
-		{{#if collapsed}}
-			<span class="collapse-switch icon-right-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
-		{{else}}
-			<span class="collapse-switch icon-down-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
+		{{> collapseArrow collapsedMedia=collapsedMediaVar}} <br>
+		{{#unless collapsed}}
 			{{#if loadImage}}
 				<figure>
 					<div class="inline-image" style="background-image: url('{{url}}');">
@@ -16,6 +14,6 @@
 					<div>click to load</div>
 				</div>
 			{{/if}}
-		{{/if}}
+		{{/unless}}
 	{{/if}}
 </template>
