+++ app/oembed/client/oembedVideoWidget.html
@@ -3,15 +3,13 @@
 	<blockquote class="message-oembed background-transparent-darker-before">
 		<div><a href="{{url}}" rel="noopener noreferrer" target="_blank">{{parsedUrl.host}}</a></div>
 		<span>{{title}}</span>
-		{{#if collapsed}}
-			<span class="collapse-switch icon-right-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
-		{{else}}
-			<span class="collapse-switch icon-down-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span><br/>
+		{{> collapseArrow collapsedMedia=collapsedMediaVar}} <br>
+		{{#unless collapsed}}
 			<video controls class="inline-video">
 				<source src="{{url}}" rel="noopener noreferrer" type="{{contentType}}">
 				Your browser does not support the video element.
 			</video>
-		{{/if}}
+		{{/unless}}
 	</blockquote>
 {{/if}}
 </template>
