--- app/oembed/client/oembedYoutubeWidget.js
@@ -1,7 +1,0 @@
-import { Template } from 'meteor/templating';
-
-Template.oembedYoutubeWidget.helpers({
-	collapsed() {
-		return this.collapsedMedia;
-	},
-});
