+++ app/ui-utils/client/lib/collapseArrow.html
@@ -1,0 +1,7 @@
+<template name='collapseArrow'>
+    {{#if collapsed}}
+        <span class="collapse-switch icon-right-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
+    {{else}}
+        <span class="collapse-switch icon-down-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
+    {{/if}}
+</template>
